tests/wpt/tests/css/css-scroll-snap/input/mouse-wheel.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#scroll-snap-type" />
<title>Mouse-wheel scroll snapping speed</title>
<meta name="assert"
      content="Test passes if mousewheel snaps without pausing">
<style>
  #scroller {
    scroll-snap-type: block mandatory;
    overflow: scroll;
    height:  400px;
    width: 400px
  }
  #space {
    width: 200px;
    height: 4000px;
  }
  .box {
    scroll-snap-align: start;
    background: blue;
    margin-bottom:  10px;
    width: 100px;
    height: 100px;
  }
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/dom/events/scrolling/scroll_support.js"></script>
<script src="../support/common.js"></script>

<div id="scroller">
  <div class="box"></div>
  <div id="target" class="box"></div>
  <div id="space"></div>
</div>
<script>
promise_test(async t => {
  await waitForCompositorReady();
  const scroller = document.getElementById("scroller");
  scroller.scrollTo(0, 0);
  assert_equals(scroller.scrollTop, 0, "verify test pre-condition");
  const scrollTop = () => {
    return scroller.scrollTop;
  };
  const scrollPromise = waitForScrollEvent(scroller);
  const wheelPromise = waitForWheelEvent(scroller);
  const targetBounds =
      document.getElementById('target').getBoundingClientRect();
  const dy = targetBounds.top / 2 + 1;
  const actions = new test_driver.Actions()
       .scroll(50, 50, 0, dy, {origin: scroller, duration: 100});
  await actions.send();
  await wheelPromise;
  await scrollPromise;
  // Detect first pause in scrolling. As expected to snap right at the end of
  // the scroll, there should be no appreciable pause before the snap takes
  // place. Once the scrolling settles, we are expected to be at the snap
  // position.
  const scrollStabilizedPromise = waitForAnimationEnd(scrollTop);
  await scrollStabilizedPromise;
  assert_approx_equals(scroller.scrollTop, 110, 0.5,
                'Failed to advance to next snap target');
}, "Wheel-scroll triggers snap to target position without intermediate pause.");
</script>
